feat: finished hmac webhook verification and security config
