+++ api/appointmentService.yaml
@@ -487,7 +487,7 @@ paths:
         content:
           application/json:
             schema:
-              $ref: '#/components/schemas/CalcomWebhookInput'
+              type: string
         required: true
       responses:
         '200':
@@ -496,8 +496,6 @@ paths:
             text/plain:
               schema:
                 type: string
-      security:
-        - Bearer: [ ]
   /processBooking:
     post:
       tags:
@@ -515,7 +513,7 @@ paths:
         content:
           application/json:
             schema:
-              $ref: '#/components/schemas/CalcomWebhookInput'
+              type: string
         required: true
       responses:
         '200':
@@ -524,8 +522,6 @@ paths:
             text/plain:
               schema:
                 type: string
-      security:
-        - Bearer: [ ]
 components:
   schemas:
     CalcomBooking:
