+++ src/main/java/com/vi/appointmentservice/api/controller/ConsultantController.java
@@ -8,6 +8,7 @@ import com.vi.appointmentservice.api.model.CalcomUser;
 import com.vi.appointmentservice.api.model.ConsultantDTO;
 import com.vi.appointmentservice.api.model.MeetingSlug;
 import com.vi.appointmentservice.generated.api.controller.ConsultantsApi;
+import com.vi.appointmentservice.helper.AuthenticatedUser;
 import io.swagger.annotations.Api;
 import java.util.List;
 import java.util.Objects;
@@ -16,6 +17,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -28,6 +30,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class ConsultantController implements ConsultantsApi {
 
+  private final @NonNull AuthenticatedUser authenticatedUser;
   private final @NonNull ConsultantFacade consultantFacade;
 
   @GetMapping(value = "/consultants/initialize", produces = {"application/json"})
@@ -63,22 +66,27 @@ public class ConsultantController implements ConsultantsApi {
   @Override
   public ResponseEntity<List<CalcomBooking>> getAllBookingsOfConsultant(String consultantId,
       String status) {
-    List<CalcomBooking> bookings;
-    if ("ACTIVE".equals(status)) {
-      bookings = consultantFacade.getConsultantActiveBookings(consultantId);
-    } else if ("EXPIRED".equals(status)) {
-      bookings = consultantFacade.getConsultantExpiredBookings(consultantId);
-    } else if ("CANCELLED".equals(status)) {
-      bookings = consultantFacade.getConsultantCancelledBookings(consultantId);
+    if (authenticatedUser.getUserId().equals(consultantId)) {
+      List<CalcomBooking> bookings;
+      if ("ACTIVE".equals(status)) {
+        bookings = consultantFacade.getConsultantActiveBookings(consultantId);
+      } else if ("EXPIRED".equals(status)) {
+        bookings = consultantFacade.getConsultantExpiredBookings(consultantId);
+      } else if ("CANCELLED".equals(status)) {
+        bookings = consultantFacade.getConsultantCancelledBookings(consultantId);
+      } else {
+        throw new BadRequestException("Given status must be ACTIVE, EXPIRED or CANCELLED");
+      }
+      return new ResponseEntity<>(bookings,
+          HttpStatus.OK);
     } else {
-      throw new BadRequestException("Given status must be ACTIVE, EXPIRED or CANCELLED");
+      throw new BadRequestException("Not authorized for given consultantId");
     }
-    return new ResponseEntity<>(bookings,
-        HttpStatus.OK);
   }
 
   @Override
-  public ResponseEntity<List<CalcomEventTypeDTO>> getAllEventTypesOfConsultant(String consultantId) {
+  public ResponseEntity<List<CalcomEventTypeDTO>> getAllEventTypesOfConsultant(
+      String consultantId) {
     return new ResponseEntity<>(consultantFacade.getAllEventTypesOfConsultantHandler(consultantId),
         HttpStatus.OK);
   }
